+++ insta/src/env.rs
@@ -25,7 +25,7 @@ pub fn get_tool_config(manifest_dir: &str) -> Arc<ToolConfig> {
 
 /// The test runner to use.
 #[cfg(feature = "_cargo_insta_internal")]
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, clap::ValueEnum)]
 pub enum TestRunner {
     Auto,
     CargoTest,
@@ -46,8 +46,8 @@ pub enum OutputBehavior {
 }
 
 /// Unreferenced snapshots flag
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg(feature = "_cargo_insta_internal")]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, clap::ValueEnum)]
 pub enum UnreferencedSnapshots {
     Auto,
     Reject,
@@ -136,7 +136,7 @@ impl ToolConfig {
                     break;
                 }
                 // ideally we would not swallow all errors here but unfortunately there are
-                // some cases where we cannot detect the error propertly.
+                // some cases where we cannot detect the error properly.
                 // Eg we can see NotADirectory here as kind, but on stable rust it cannot
                 // be matched on.
                 Err(_) => continue,