+++ insta/src/snapshot.rs
@@ -133,7 +133,8 @@ impl PendingInlineSnapshot {
 pub struct MetaData {
     /// The source file (relative to workspace root).
     pub(crate) source: Option<String>,
-    /// The source line if available.
+    /// The source line, if available. This is used by pending snapshots, but trimmed
+    /// before writing to the final `.snap` files in [`MetaData::trim_for_persistence`].
     pub(crate) assertion_line: Option<u32>,
     /// Optional human readable (non formatted) snapshot description.
     pub(crate) description: Option<String>,
@@ -246,8 +247,15 @@ impl MetaData {
         Content::Struct("MetaData", fields)
     }
 
-    /// Trims the metadata for persistence.
+    /// Trims the metadata of fields that we don't save to `.snap` files (those
+    /// we only use for display while reviewing)
     fn trim_for_persistence(&self) -> Cow<'_, MetaData> {
+        // TODO: in order for `--require-full-match` to work on inline snapshots
+        // without cargo-insta, we need to trim all fields if there's an inline
+        // snapshot. But we don't know that from here (notably
+        // `self.input_file.is_none()` is not a correct approach). Given that
+        // `--require-full-match` is experimental and we're working on making
+        // inline & file snapshots more coherent, I'm leaving this as is for now.
         if self.assertion_line.is_some() {
             let mut rv = self.clone();
             rv.assertion_line = None;
@@ -311,6 +319,7 @@ impl Snapshot {
                     }
                 }
             }
+            eprintln!("A snapshot uses an old snapshot format; please update it to the new format with `cargo insta --force-update-snapshots.\n\nSnapshot is at: {}", p.to_string_lossy());
             rv
         };
 
@@ -323,31 +332,11 @@ impl Snapshot {
             buf.push_str(&line);
         }
 
-        let module_name = p
-            .file_name()
-            .unwrap()
-            .to_str()
-            .unwrap_or("")
-            .split("__")
-            .next()
-            .unwrap_or("<unknown>")
-            .to_string();
-
-        let snapshot_name = p
-            .file_name()
-            .unwrap()
-            .to_str()
-            .unwrap_or("")
-            .split('.')
-            .next()
-            .unwrap_or("")
-            .splitn(2, "__")
-            .nth(1)
-            .map(|x| x.to_string());
+        let (snapshot_name, module_name) = names_of_path(p);
 
         Ok(Snapshot::from_components(
             module_name,
-            snapshot_name,
+            Some(snapshot_name),
             metadata,
             buf.into(),
         ))
@@ -442,7 +431,8 @@ impl Snapshot {
 
     /// Snapshot contents _and_ metadata match another snapshot's.
     pub fn matches_fully(&self, other: &Snapshot) -> bool {
-        self.matches(other) && self.metadata == other.metadata
+        self.matches(other)
+            && self.metadata.trim_for_persistence() == other.metadata.trim_for_persistence()
     }
 
     /// The snapshot contents as a &str
@@ -471,11 +461,19 @@ impl Snapshot {
         let serialized_snapshot = self.serialize_snapshot(md);
 
         // check the reference file for contents.  Note that we always want to
-        // compare snapshots that were trimmed to persistence here.
+        // compare snapshots that were trimmed to persistence here.  This is a
+        // stricter check than even `matches_fully`, since it's comparing the
+        // exact contents of the file.
         if let Ok(old) = fs::read_to_string(ref_file.unwrap_or(path)) {
             let persisted = match md.trim_for_persistence() {
                 Cow::Owned(trimmed) => Cow::Owned(self.serialize_snapshot(&trimmed)),
-                Cow::Borrowed(_) => Cow::Borrowed(&serialized_snapshot),
+                Cow::Borrowed(trimmed) => {
+                    // This condition needs to hold, otherwise we need to
+                    // compare the old value to a newly trimmed serialized snapshot
+                    debug_assert_eq!(trimmed, md);
+
+                    Cow::Borrowed(&serialized_snapshot)
+                }
             };
             if old == persisted.as_str() {
                 return Ok(false);
@@ -501,8 +499,7 @@ impl Snapshot {
     /// If the existing snapshot matches the new file, then `None` is returned, otherwise
     /// the name of the new snapshot file.
     pub(crate) fn save_new(&self, path: &Path) -> Result<Option<PathBuf>, Box<dyn Error>> {
-        let mut new_path = path.to_path_buf();
-        new_path.set_extension("snap.new");
+        let new_path = path.to_path_buf().with_extension("snap.new");
         if self.save_with_metadata(&new_path, Some(path), &self.metadata)? {
             Ok(Some(new_path))
         } else {
@@ -523,17 +520,31 @@ impl SnapshotContents {
 
     /// Returns the snapshot contents as string with surrounding whitespace removed.
     pub fn as_str(&self) -> &str {
-        self.0
-            .trim_start_matches(|x| x == '\r' || x == '\n')
-            .trim_end()
+        self.0.trim_start_matches(['\r', '\n']).trim_end()
     }
 
     pub fn to_inline(&self, indentation: usize) -> String {
         let contents = &self.0;
         let mut out = String::new();
-        let is_escape = contents.contains(&['\n', '\\', '"'][..]);
 
-        out.push_str(if is_escape { "r###\"" } else { "\"" });
+        // Escape the string if needed, with `r#`, using with 1 more `#` than
+        // the maximum number of existing contiguous `#`.
+        let is_escape = contents.contains(&['\n', '\\', '"'][..]);
+        let delimiter = if is_escape {
+            let max_contiguous_hash = contents
+                .split(|c| c != '#')
+                .map(|group| group.len())
+                .max()
+                .unwrap_or(0);
+            out.push('r');
+            "#".repeat(max_contiguous_hash + 1)
+        } else {
+            "".to_string()
+        };
+
+        out.push_str(&delimiter);
+        out.push('"');
+
         // if we have more than one line we want to change into the block
         // representation mode
         if contents.contains('\n') {
@@ -557,7 +568,8 @@ impl SnapshotContents {
             out.push_str(contents);
         }
 
-        out.push_str(if is_escape { "\"###" } else { "\"" });
+        out.push('"');
+        out.push_str(&delimiter);
 
         out
     }
@@ -630,6 +642,53 @@ fn normalize_inline_snapshot(snapshot: &str) -> String {
         .join("\n")
 }
 
+/// Extracts the module and snapshot name from a snapshot path
+fn names_of_path(path: &Path) -> (String, String) {
+    // The final part of the snapshot file name is the test name; the
+    // initial parts are the module name
+    let parts: Vec<&str> = path
+        .file_stem()
+        .unwrap()
+        .to_str()
+        .unwrap_or("")
+        .rsplitn(2, "__")
+        .collect();
+
+    match parts.as_slice() {
+        [snapshot_name, module_name] => (snapshot_name.to_string(), module_name.to_string()),
+        [snapshot_name] => (snapshot_name.to_string(), String::new()),
+        _ => (String::new(), "<unknown>".to_string()),
+    }
+}
+
+#[test]
+fn test_names_of_path() {
+    assert_debug_snapshot!(
+        names_of_path(Path::new("/src/snapshots/insta_tests__tests__name_foo.snap")), @r###"
+    (
+        "name_foo",
+        "insta_tests__tests",
+    )
+    "###
+    );
+    assert_debug_snapshot!(
+        names_of_path(Path::new("/src/snapshots/name_foo.snap")), @r###"
+    (
+        "name_foo",
+        "",
+    )
+    "###
+    );
+    assert_debug_snapshot!(
+        names_of_path(Path::new("foo/src/snapshots/go1.20.5.snap")), @r###"
+    (
+        "go1.20.5",
+        "",
+    )
+    "###
+    );
+}
+
 /// Helper function that returns the real inline snapshot value from a given
 /// frozen value string.  If the string starts with the '⋮' character
 /// (optionally prefixed by whitespace) the alternative serialization format
@@ -641,6 +700,8 @@ fn get_inline_snapshot_value(frozen_value: &str) -> String {
     // (the only call site)
 
     if frozen_value.trim_start().starts_with('⋮') {
+        eprintln!("A snapshot uses an old snapshot format; please update it to the new format with `cargo insta --force-update-snapshots.\n\nValue: {}", frozen_value);
+
         // legacy format - retain so old snapshots still work
         let mut buf = String::new();
         let mut line_iter = frozen_value.lines();
@@ -693,10 +754,10 @@ a
 b"[1..];
     assert_eq!(
         SnapshotContents(t.to_string()).to_inline(0),
-        "r###\"
+        "r#\"
 a
 b
-\"###"
+\"#"
     );
 
     let t = &"
@@ -704,10 +765,10 @@ a
 b"[1..];
     assert_eq!(
         SnapshotContents(t.to_string()).to_inline(4),
-        "r###\"
+        "r#\"
     a
     b
-    \"###"
+    \"#"
     );
 
     let t = &"
@@ -715,23 +776,23 @@ b"[1..];
     b"[1..];
     assert_eq!(
         SnapshotContents(t.to_string()).to_inline(0),
-        "r###\"
+        "r#\"
     a
     b
-\"###"
+\"#"
     );
 
     let t = &"
-    a
+a
 
-    b"[1..];
+b"[1..];
     assert_eq!(
-        SnapshotContents(t.to_string()).to_inline(0),
-        "r###\"
+        SnapshotContents(t.to_string()).to_inline(4),
+        "r#\"
     a
 
     b
-\"###"
+    \"#"
     );
 
     let t = &"
@@ -739,13 +800,28 @@ b"[1..];
 "[1..];
     assert_eq!(
         SnapshotContents(t.to_string()).to_inline(0),
-        "r###\"
+        "r#\"
     ab
-\"###"
+\"#"
     );
 
     let t = "ab";
     assert_eq!(SnapshotContents(t.to_string()).to_inline(0), r#""ab""#);
+}
+
+#[test]
+fn test_snapshot_contents_hashes() {
+    let t = "a###b";
+    assert_eq!(SnapshotContents(t.to_string()).to_inline(0), r#""a###b""#);
+
+    let t = "a\n###b";
+    assert_eq!(
+        SnapshotContents(t.to_string()).to_inline(0),
+        r#####"r####"
+a
+###b
+"####"#####
+    );
 }
 
 #[test]
@@ -915,7 +991,7 @@ fn test_parse_yaml_error() {
 
     let error = format!("{}", Snapshot::from_file(temp.as_path()).unwrap_err());
     assert!(error.contains("Failed parsing the YAML from"));
-    assert!(error.contains("/bad.yaml"));
+    assert!(error.contains("bad.yaml"));
 }
 
 /// Check that snapshots don't take ownership of the value